+++ .eslintrc.yml
@@ -1,3 +1,6 @@
+parserOptions:
+    ecmaVersion: 2015
+    sourceType: 'module'
 env:
   es6: true
   browser: true
