+++ public/index.html
@@ -32,6 +32,6 @@
     <!-- Количество фильмов -->
   </section>
 </footer>
-<script src="../src/main.js"></script>
+<script src="../bundle.js"></script>
 </body>
 </html>
