+++ src/components/stat.js
@@ -1,0 +1,5 @@
+export const createStatTemplate = () => {
+  return (`
+  <p>130 291 movies inside</p>
+  `);
+};
