+++ src/const.js
@@ -1,0 +1,5 @@
+export const countsFilmCards = {
+  MAIN: 5,
+  TOP_RATED: 2,
+  MOST_COMMENTED: 2
+};
