+++ src/utils.js
@@ -1,0 +1,3 @@
+export const render = (container, template, place = `beforeend`) => {
+  container.insertAdjacentHTML(place, template);
+};
